A bug in computing the ANF for 3-bit and 4-bit S-boxes
In the function `get_coordinates_ANF()`, for 3-bit and 4-bit S-boxes, at the beginning, to obtain the truth table, the mask should be AND to the value, not XOR.

The influence of this bug:
For 3-bit and 4-bit S-boxes, the computed ANF of coordinates are all XORed with '1', which should not!

There is no influence on other results and on 5/6/7/8-bit S-boxes.